# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the user service for the Byow app",
        "title": "Build Your Own Website User Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/companies/all": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Find All Companies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Keyword",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CompanyListResponseSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/companies/create": {
            "post": {
                "description": "Register a new company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Create Company",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Cemerlang Jaya",
                        "description": "Company Name",
                        "name": "company_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"john@company.com\"",
                        "description": "Company Email",
                        "name": "company_email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "628112123123",
                        "description": "Company Phone",
                        "name": "company_phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"123 Cemerlang St, Tech City\"",
                        "description": "Company Address",
                        "name": "company_address",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Company Logo",
                        "name": "company_logo",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CompanyRequestSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/companies/{id}": {
            "get": {
                "description": "Get company details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Get Company By ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"60d5ec49f1c2b14c88f3c5e5\"",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CompanyRequestSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Update Company",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"60d5ec49f1c2b14c88f3c5e5\"",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Cemerlang Jaya",
                        "description": "Company Name",
                        "name": "company_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "example": "\"john@company.com\"",
                        "description": "Company Email",
                        "name": "company_email",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "example": "628112123123",
                        "description": "Company Phone",
                        "name": "company_phone",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "example": "\"123 Cemerlang St, Tech City\"",
                        "description": "Company Address",
                        "name": "company_address",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Company Logo",
                        "name": "company_logo",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CompanyRequestSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Company by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Delete Company",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"60d5ec49f1c2b14c88f3c5e5\"",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CompanyRequestSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CompanyListResponseSwagger": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CompanyResponse"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "SUCCESS"
                }
            }
        },
        "dto.CompanyRequestSwagger": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/dto.CompanyResponse"
                },
                "status": {
                    "type": "string",
                    "example": "SUCCESS"
                }
            }
        },
        "dto.CompanyResponse": {
            "type": "object",
            "properties": {
                "company_address": {
                    "type": "string",
                    "example": "123 BuildYow St, Tech City"
                },
                "company_email": {
                    "type": "string",
                    "example": "info@buildyow.com"
                },
                "company_id": {
                    "type": "string",
                    "example": "60c72b2f9b1e8c001c8e4d3a"
                },
                "company_logo": {
                    "type": "string",
                    "example": "https://assets/images/company_logo.jpg"
                },
                "company_name": {
                    "type": "string",
                    "example": "BuildYow"
                },
                "company_phone": {
                    "type": "string",
                    "example": "628112123123"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-10-01T12:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "60c72b2f9b1e8c001c8e4d3a"
                },
                "verified": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {},
                "message": {
                    "type": "string",
                    "example": "Validation failed"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {
                    "$ref": "#/definitions/dto.ErrorResponseData"
                },
                "error": {
                    "$ref": "#/definitions/dto.ErrorDetail"
                },
                "status": {
                    "type": "string",
                    "example": "ERROR"
                }
            }
        },
        "dto.ErrorResponseData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "INTERNAL_SERVER_ERROR"
                }
            }
        },
...
    }
}